+++ unicode-data-security/test/Unicode/Char/Identifiers/SecuritySpec.hs
@@ -6,6 +6,7 @@ module Unicode.Char.Identifiers.SecuritySpec
 
 import Control.Monad (when)
 import Data.Foldable (traverse_)
+import qualified Data.List.NonEmpty as NE
 import Unicode.Char.Identifiers.Security
 import Test.Hspec
 
@@ -13,37 +14,38 @@ spec :: Spec
 spec = do
     describe "identifierTypes" do
         it "some values" do
-            let check ts = (ts ==) . identifierTypes
-            minBound  `shouldSatisfy` check [NotCharacter]
-            maxBound  `shouldSatisfy` check [NotCharacter]
-            '1'       `shouldSatisfy` check [Recommended]
-            'a'       `shouldSatisfy` check [Recommended]
-            '\x0149'  `shouldSatisfy` check [Deprecated]
-            '\x00AD'  `shouldSatisfy` check [DefaultIgnorable]
-            '\x1F240' `shouldSatisfy` check [NotNFKC]
-            '\x1F4FF' `shouldSatisfy` check [NotXID]
-            '\x082C'  `shouldSatisfy` check [Exclusion]
-            '\x1680'  `shouldSatisfy` check [Exclusion, NotXID]
-            '\x01B9'  `shouldSatisfy` check [Obsolete]
-            '\x05C6'  `shouldSatisfy` check [Obsolete, NotXID]
-            '\x0234'  `shouldSatisfy` check [Technical]
-            '\x2CF1'  `shouldSatisfy` check [Technical, Exclusion]
-            '\x20E4'  `shouldSatisfy` check [Technical, NotXID]
-            '\x018D'  `shouldSatisfy` check [Technical, Obsolete]
-            '\x2E00'  `shouldSatisfy` check [Technical, Obsolete, NotXID]
-            '\x018E'  `shouldSatisfy` check [UncommonUse]
-            '\x16A40' `shouldSatisfy` check [UncommonUse, Exclusion]
-            '\x05A2'  `shouldSatisfy` check [UncommonUse, Obsolete]
-            '\xA8FC'  `shouldSatisfy` check [UncommonUse, Obsolete, NotXID]
-            '\x025B'  `shouldSatisfy` check [UncommonUse, Technical]
-            '\x1D1DE' `shouldSatisfy` check [UncommonUse, Technical, NotXID]
-            '\x07FD'  `shouldSatisfy` check [LimitedUse]
-            '\xA9CF'  `shouldSatisfy` check [LimitedUse, Exclusion]
-            '\x070F'  `shouldSatisfy` check [LimitedUse, NotXID]
-            '\x07FA'  `shouldSatisfy` check [LimitedUse, Obsolete]
-            '\x1B6B'  `shouldSatisfy` check [LimitedUse, Technical]
-            '\x2019'  `shouldSatisfy` check [Inclusion]
-            '\x018F'  `shouldSatisfy` check [Recommended]
+            let shouldBe' :: (HasCallStack) => Char -> NE.NonEmpty IdentifierType -> Expectation
+                shouldBe' c = shouldBe (identifierTypes c)
+            minBound  `shouldBe'` [NotCharacter]
+            maxBound  `shouldBe'` [NotCharacter]
+            '1'       `shouldBe'` [Recommended]
+            'a'       `shouldBe'` [Recommended]
+            '\x0149'  `shouldBe'` [Deprecated]
+            '\x00AD'  `shouldBe'` [DefaultIgnorable]
+            '\x1F240' `shouldBe'` [NotNFKC]
+            '\x1F4FF' `shouldBe'` [NotXID]
+            '\x082C'  `shouldBe'` [Exclusion]
+            '\x1680'  `shouldBe'` [Exclusion, NotXID]
+            '\x01B9'  `shouldBe'` [Obsolete]
+            '\x05C6'  `shouldBe'` [Obsolete, NotXID]
+            '\x0234'  `shouldBe'` [Technical]
+            '\x2CF1'  `shouldBe'` [Technical, Exclusion]
+            '\x20E4'  `shouldBe'` [Technical, NotXID]
+            '\x018D'  `shouldBe'` [Technical, Obsolete]
+            '\x2E00'  `shouldBe'` [Technical, Obsolete, NotXID]
+            '\x018E'  `shouldBe'` [UncommonUse]
+            '\x16A40' `shouldBe'` [UncommonUse, Exclusion]
+            '\x05A2'  `shouldBe'` [UncommonUse, Obsolete]
+            '\xA8FC'  `shouldBe'` [UncommonUse, Obsolete, NotXID]
+            '\x025B'  `shouldBe'` [UncommonUse, Technical]
+            '\x1D1DE' `shouldBe'` [UncommonUse, Technical, NotXID]
+            '\x07FD'  `shouldBe'` [LimitedUse]
+            '\xA9CF'  `shouldBe'` [LimitedUse, UncommonUse]
+            '\x070F'  `shouldBe'` [LimitedUse, NotXID]
+            '\x07FA'  `shouldBe'` [LimitedUse, Obsolete]
+            '\x1B6B'  `shouldBe'` [LimitedUse, Technical]
+            '\x2019'  `shouldBe'` [Inclusion]
+            '\x018F'  `shouldBe'` [Recommended]
         it "invariants" do
             let {
                 check c = do